Refactor AWS workflow and playbook
This commit refactors the AWS workflow and playbook files. It removes the caching of the hosts file and updates the configuration and deployment steps. Additionally, it changes the remote user from "ec2-user" to "ubuntu" in the playbook. The package management commands have also been updated from "yum" to "apt".